requirements.txt
azure-functions
azure-servicebus
azure-storage-blob==12.6.0
requests